EMS-157 - fix e2e test issue with check-answers where old session data would persist
