+++ e2e-tests/cypress/integration/journeys/check-your-answers.spec.js
@@ -47,7 +47,6 @@ context('Check your answers page', () => {
 
   beforeEach(() => {
     Cypress.Cookies.preserveOnce('_csrf');
-    Cypress.Cookies.preserveOnce('connect.sid');
   });
 
   it('passes the audits', () => {
